Split the decryption routines out of main, into their own decrypt file.
