+++ source/main.c
@@ -2,7 +2,7 @@
 #include <stdlib.h>
 #include <stdio.h>
 
-#include "crypto.h"
+#include "decrypt.h"
 #include "draw.h"
 #include "hid.h"
 #include "ncch.h"
@@ -10,184 +10,9 @@
 #include "gamecart/protocol.h"
 #include "gamecart/command_ctr.h"
 
-extern s32 CartID;
-extern s32 CartID2;
-
 // File IO utility functions
 static FATFS fs;
 static FIL file;
-
-enum NCCH_Type {
-    NCCHTYPE_EXHEADER,
-    NCCHTYPE_EXEFS,
-    NCCHTYPE_ROMFS
-};
-
-static void ClearTop(void) {
-    ClearScreen(TOP_SCREEN0, RGB(255, 255, 255));
-    ClearScreen(TOP_SCREEN1, RGB(255, 255, 255));
-    current_y = 0;
-}
-
-static void wait_key(void) {
-    Debug("Press key to continue...");
-    InputWait();
-}
-
-static void ncch_get_counter(NCCH_Header* header, u8 counter[16], enum NCCH_Type type) {
-    u32 version = header->version;
-    u32 mediaunitsize = 0x200; // TODO
-    u8* partitionid = header->partition_id;
-    u32 x = 0;
-
-    memset(counter, 0, 16);
-
-    switch (version) {
-    case 0:
-    case 2:
-        for(u32 i = 0; i < 8; i++)
-            counter[i] = partitionid[7 - i];
-        counter[8] = type;
-        break;
-    case 1:
-        switch (type) {
-        case NCCHTYPE_EXHEADER:
-            x = 0x200;
-            break;
-        case NCCHTYPE_EXEFS:
-            x = header->exefs_offset * mediaunitsize;
-            break;
-        case NCCHTYPE_ROMFS:
-            x = header->romfs_offset * mediaunitsize;
-            break;
-        }
-
-        for(u32 i = 0; i < 8; i++)
-            counter[i] = partitionid[i];
-        for(u32 i = 0; i < 4; i++)
-            counter[12 + i] = x >> ((3 - i) * 8);
-        break;
-    }
-}
-
-// Generate list of encrypted data blocks, ordered by file offset
-typedef struct {
-    uint32_t addr; // mediaUnits
-    uint32_t size; // mediaUnits
-    uint32_t uses_7x_crypto;
-    uint32_t pad;
-    u8 ctr[16] __attribute__((aligned(16)));
-    u8 keyY[16] __attribute__((aligned(16)));
-} EncryptedArea;
-
-struct Context {
-    u32 decrypt;
-
-    u8* buffer;
-    size_t buffer_size;
-
-    u32 cart_size;
-    u32 media_unit;
-
-    EncryptedArea areas[16];
-    int num_areas;
-
-    NCCH_Header ncchs[8]; // one ncch per partition
-
-};
-
-static int find_encrypted_ncch_regions(NCCH_Header* ncch, NCSD_Header* ncsd, unsigned partition, struct Context* ctx, u32 initial_sector, u32 dumped) {
-    u8* source = ctx->buffer;
-    u8* dest = (u8*)ncch;
-    u32 size = sizeof(NCCH_Header);
-    Debug("Detected NCCH: %p, %p, %x, %x", source, dest, size);
-    if (ncsd->partition_table[partition].offset > initial_sector) {
-        u32 delta = (ncsd->partition_table[partition].offset - initial_sector) * ctx->media_unit;
-        source += delta;
-    }
-    if (ncsd->partition_table[partition].offset < initial_sector) {
-        // TODO: I don't think this can ever happen. If it does happen, I don't think it works correctly.
-        u32 delta = (initial_sector - ncsd->partition_table[partition].offset) * ctx->media_unit;
-        dest += delta;
-        size -= delta;
-    }
-    // TODO: Should make more special cases, I think.
-    memcpy(dest, source, size);
-
-    // If the header is complete
-    if (initial_sector + dumped >= ncsd->partition_table[partition].offset + sizeof(NCCH_Header)) {
-        if (ncch->flags[3] != 0) {
-            // This cartridge uses the new crypto method introduced in system version 7.x.
-            // TODO: Set the uses_7x_crypto field of the affected encrypted areas appropriately.
-            // TODO: System versions older than 7.x require the keyX to be given by the user. If the key is not given, we should abort dumping.
-            Debug("ERROR: Cartridge uses 7.x crypto.");
-            Debug("uncart does not support this, yet.");
-            Debug("Perform a raw dump and decrypt it with");
-            Debug("an external tool.");
-            Debug("Dumping process aborted.");
-            wait_key();
-            return -1;
-        }
-
-        // TODO: Consider checking for other flags, in particular the 9.6.0 keyY generator
-
-        if (ncch->extended_header_size) {
-            ctx->areas[ctx->num_areas].addr = ncsd->partition_table[partition].offset + sizeof(NCCH_Header) / ctx->media_unit;
-            ctx->areas[ctx->num_areas].size = 0x800 / ctx->media_unit; // NOTE: ncchs->extended_header_size doesn't cover the full exheader!
-            ctx->areas[ctx->num_areas].uses_7x_crypto = 0;
-            memcpy(ctx->areas[ctx->num_areas].keyY, ncch->signature, sizeof(ctx->areas[ctx->num_areas].keyY));
-            ncch_get_counter(ncch, ctx->areas[ctx->num_areas].ctr, NCCHTYPE_EXHEADER);
-            ctx->num_areas++;
-        }
-
-        if (ncch->exefs_offset && ncch->exefs_size) {
-            ctx->areas[ctx->num_areas].addr = ncsd->partition_table[partition].offset + ncch->exefs_offset;
-            ctx->areas[ctx->num_areas].size = ncch->exefs_size;
-            memcpy(ctx->areas[ctx->num_areas].keyY, ncch->signature, sizeof(ctx->areas[ctx->num_areas].keyY));
-            ctx->areas[ctx->num_areas].uses_7x_crypto = 0;
-            ncch_get_counter(ncch, ctx->areas[ctx->num_areas].ctr, NCCHTYPE_EXEFS);
-            ctx->num_areas++;
-        }
-        if (ncch->romfs_offset && ncch->romfs_size) {
-            ctx->areas[ctx->num_areas].addr = ncsd->partition_table[partition].offset + ncch->romfs_offset;
-            ctx->areas[ctx->num_areas].uses_7x_crypto = 0;
-            ctx->areas[ctx->num_areas].size = ncch->romfs_size;
-            memcpy(ctx->areas[ctx->num_areas].keyY, ncch->signature, sizeof(ctx->areas[ctx->num_areas].keyY));
-            ncch_get_counter(ncch, ctx->areas[ctx->num_areas].ctr, NCCHTYPE_ROMFS);
-            ctx->num_areas++;
-        }
-    }
-    return 0;
-}
-
-static void decrypt_region(EncryptedArea area, struct Context* ctx, u32 initial_sector, u32 dumped) {
-    u32 keyslot = (area.uses_7x_crypto == 0xA) ? 0x18 : (area.uses_7x_crypto != 0) ? 0x25 : 0x2c;
-    setup_aeskey(keyslot, AES_BIG_INPUT | AES_NORMAL_INPUT, area.keyY);
-    use_aeskey(keyslot);
-
-    static const uint8_t zero_buf[16] __attribute__((aligned(16))) = {0};
-    static const uint8_t dec_buf[16] __attribute__((aligned(16))) = {0};
-    u32 adr2 = (initial_sector < area.addr) ? area.addr : initial_sector;
-    u32 limit = initial_sector + dumped;
-    if (initial_sector + dumped > area.addr + area.size)
-        limit = area.addr + area.size;
-    adr2 *= ctx->media_unit / 16;
-    limit *= ctx->media_unit / 16;
-
-    Debug("Decrypting from %08x to %08x", adr2 * 16, limit * 16);
-
-    while (adr2 < limit) {
-        set_ctr(AES_BIG_INPUT | AES_NORMAL_INPUT, area.ctr);
-        aes_decrypt((void*)zero_buf, (void*)dec_buf, area.ctr, 1, AES_CTR_MODE);
-        add_ctr(area.ctr, 1);
-
-        u8* dest = ctx->buffer + adr2 * 16 - (initial_sector * ctx->media_unit);
-        for (int k = 0; k < 16; ++k) {
-            dest[k] ^= dec_buf[k];
-        }
-        adr2++;
-    }
-}
 
 int dump_cart_region(u32 start_sector, u32 end_sector, FIL* output_file, struct Context* ctx, NCSD_Header* ncsd) {
     const u32 read_size = 1 * 1024 * 1024 / ctx->media_unit; // 1MB
@@ -260,7 +85,7 @@ restart_program:
     ClearTop();
     Debug("ROM dump tool v0.2");
     Debug("Insert your game cart now.");
-    wait_key();
+    WaitKey();
 
     // Arbitrary target buffer
     // TODO: This should be done in a nicer way ;)
@@ -352,13 +177,13 @@ restart_program:
 
         if (f_mount(&fs, "0:", 0) != FR_OK) {
             Debug("Failed to f_mount... Retrying");
-            wait_key();
+            WaitKey();
             goto cleanup_none;
         }
 
         if (f_open(&file, filename_buf, FA_READ | FA_WRITE | FA_CREATE_ALWAYS) != FR_OK) {
             Debug("Failed to create file... Retrying");
-            wait_key();
+            WaitKey();
             goto cleanup_mount;
         }
